pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "goodmem-adk"
version = "0.1.0"
description = "Goodmem plugin for Google ADK — persistent memory for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Forrest Sheng Bao", email = "forrest@goodmem.ai" },
    { name = "Bashar Alrfooh", email = "bashar.alrfooh@goodmem.ai" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "google-adk>=1.0.0",
    "httpx",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "fpdf2",
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
include = ["goodmem_adk*"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
markers = [
    "integration: requires live Goodmem server and Gemini API key",
]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
